moved timeseries hdf5 dependencies into hdf5 connector
